created model

+++ src/main/java/com/propertydekho/PropertyDetails.java
@@ -7,5 +7,5 @@ import lombok.Data;
 public class PropertyDetails {
     @JsonProperty("prop_id") private String propID;
     @JsonProperty("prop_name") private String propName;
-    @JsonProperty("prop_price") private double propPrice;
+    @JsonProperty("prop_price") private long propPrice;
 }
